collector/requirements.txt
psutil
psycopg2-binary

